Organize test structure - Move unit tests to 'ut' package.
