+++ src/test/java/com/github/dakusui/actionunit/ut/ExamplesTest.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.actionunit;
+package com.github.dakusui.actionunit.ut;
 
 import com.github.dakusui.actionunit.examples.Basic;
 import com.github.dakusui.actionunit.examples.NameOfAction;
