+++ src/test/java/com/github/dakusui/actionunit/ut/UtilsTest.java
@@ -1,5 +1,6 @@
-package com.github.dakusui.actionunit;
+package com.github.dakusui.actionunit.ut;
 
+import com.github.dakusui.actionunit.Utils;
 import org.junit.Test;
 
 import java.util.List;
